minor code changes: removed unused or faulty used mocks, updated tests, removed weak warnings and warnings, removed logRender

+++ data/blockchain/metachain.go
@@ -100,7 +100,7 @@ func (mc *MetaChain) SetCurrentBlockHeader(header data.HeaderHandler) error {
 	}
 
 	mc.appStatusHandler.SetUInt64Value(core.MetricNonce, currHead.Nonce)
-	mc.appStatusHandler.SetUInt64Value(core.MetricSynchronizedRound, uint64(currHead.Round))
+	mc.appStatusHandler.SetUInt64Value(core.MetricSynchronizedRound, currHead.Round)
 
 	mc.CurrentBlock = currHead
